+++ include/ignition/rendering/base/BaseScene.hh
@@ -279,6 +279,21 @@ namespace ignition
                   const std::string &_name) override;
 
       // Documentation inherited.
+      public: virtual ThermalCameraPtr CreateThermalCamera() override;
+
+      // Documentation inherited.
+      public: virtual ThermalCameraPtr CreateThermalCamera(
+                  const unsigned int _id) override;
+
+      // Documentation inherited.
+      public: virtual ThermalCameraPtr CreateThermalCamera(
+                  const std::string &_name) override;
+
+      // Documentation inherited.
+      public: virtual ThermalCameraPtr CreateThermalCamera(
+                  const unsigned int _id, const std::string &_name) override;
+
+      // Documentation inherited.
       public: virtual GpuRaysPtr CreateGpuRays() override;
 
       // Documentation inherited.
@@ -359,6 +374,9 @@ namespace ignition
       public: virtual GridPtr CreateGrid() override;
 
       // Documentation inherited.
+      public: virtual MarkerPtr CreateMarker() override;
+
+      // Documentation inherited.
       public: virtual TextPtr CreateText() override;
 
       public: virtual MaterialPtr CreateMaterial(const std::string &_name = "")
@@ -405,6 +423,17 @@ namespace ignition
 
       protected: virtual DepthCameraPtr CreateDepthCameraImpl(unsigned int _id,
                      const std::string &_name) = 0;
+
+      /// \brief Implementation for creating a thermal camera.
+      /// \param[in] _id Unique id
+      /// \param[in] _name Name of thermal camera
+      protected: virtual ThermalCameraPtr CreateThermalCameraImpl(
+                     unsigned int /*_id*/, const std::string &/*_name*/)
+                 {
+                   ignerr << "Thermal camera not supported by: "
+                          << this->Engine()->Name() << std::endl;
+                   return ThermalCameraPtr();
+                 }
 
       /// \brief Implementation for creating GpuRays sensor.
       /// \param[in] _id Unique id
@@ -461,6 +490,13 @@ namespace ignition
       /// \param[in] _name unique object name.
       /// \return Pointer to a grid geometry object
       protected: virtual GridPtr CreateGridImpl(unsigned int _id,
+                     const std::string &_name) = 0;
+
+      /// \brief Implementation for creating a marker geometry object
+      /// \param[in] _id unique object id.
+      /// \param[in] _name unique object name.
+      /// \return Pointer to a marker geometry object
+      protected: virtual MarkerPtr CreateMarkerImpl(unsigned int _id,
                      const std::string &_name) = 0;
 
       /// \brief Implementation for creating a text's geometry object