+++ ogre2/include/ignition/rendering/ogre2/Ogre2RenderTypes.hh
@@ -35,6 +35,7 @@ namespace ignition
     class Ogre2GizmoVisual;
     class Ogre2GpuRays;
     class Ogre2Light;
+    class Ogre2Marker;
     class Ogre2Material;
     class Ogre2Mesh;
     class Ogre2MeshFactory;
@@ -51,6 +52,7 @@ namespace ignition
     class Ogre2Sensor;
     class Ogre2SpotLight;
     class Ogre2SubMesh;
+    class Ogre2ThermalCamera;
     class Ogre2Visual;
 
     typedef BaseGeometryStore<Ogre2Geometry>      Ogre2GeometryStore;
@@ -70,6 +72,7 @@ namespace ignition
     typedef shared_ptr<Ogre2GizmoVisual>          Ogre2GizmoVisualPtr;
     typedef shared_ptr<Ogre2GpuRays>              Ogre2GpuRaysPtr;
     typedef shared_ptr<Ogre2Light>                Ogre2LightPtr;
+    typedef shared_ptr<Ogre2Marker>               Ogre2MarkerPtr;
     typedef shared_ptr<Ogre2Material>             Ogre2MaterialPtr;
     typedef shared_ptr<Ogre2Mesh>                 Ogre2MeshPtr;
     typedef shared_ptr<Ogre2MeshFactory>          Ogre2MeshFactoryPtr;
@@ -85,6 +88,7 @@ namespace ignition
     typedef shared_ptr<Ogre2Sensor>               Ogre2SensorPtr;
     typedef shared_ptr<Ogre2SpotLight>            Ogre2SpotLightPtr;
     typedef shared_ptr<Ogre2SubMesh>              Ogre2SubMeshPtr;
+    typedef shared_ptr<Ogre2ThermalCamera>        Ogre2ThermalCameraPtr;
     typedef shared_ptr<Ogre2Visual>               Ogre2VisualPtr;
 
     typedef shared_ptr<Ogre2GeometryStore>        Ogre2GeometryStorePtr;