+++ src/PixelFormat.cc
@@ -35,7 +35,8 @@ const char *PixelUtil::names[PF_COUNT] =
       "BAYER_GRGB8",
       "FLOAT32_R",
       "FLOAT32_RGBA",
-      "FLOAT32_RGB"
+      "FLOAT32_RGB",
+      "L16"
     };
 
 //////////////////////////////////////////////////
@@ -58,11 +59,13 @@ const unsigned char PixelUtil::channelCounts[PF_COUNT] =
       // BAYER_GRGB8
       4,
       // PF_FLOAT32_R
-      4,
+      1,
       // PF_FLOAT32_RGBA
-      8,
+      4,
       // PF_FLOAT32_RGB
-      8
+      3,
+      // PF_L16
+      1
     };
 
 //////////////////////////////////////////////////
@@ -85,11 +88,13 @@ const unsigned char PixelUtil::channelByteCounts[PF_COUNT] =
       // BAYER_GRGB8
       1,
       // PF_FLOAT32_R
-      1,
+      4,
       // PF_FLOAT32_RGBA
-      1,
+      4,
       // PF_FLOAT32_RGB
-      1
+      4,
+      // PF_L16
+      2,
     };
 
 //////////////////////////////////////////////////