+++ Changelog.md
@@ -4,9 +4,67 @@
 
 ### Ignition Rendering 3.0.0 (20XX-XX-XX)
 
-### Ignition Rendering 2.X
+1. Support setting cast shadows
+    * [Pull request 209](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/209)
 
-### Ignition Rendering 2.0.0 (20XX-XX-XX)
+1. Add Thermal Camera
+    * [Pull request 202](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/202)
+
+1. Add support for emissive map
+    * [Pull request 207](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/207)
+
+1. Add Marker class and port ogre dynamic lines and renderable
+    * [Pull request 195](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/195)
+    * [Pull request 208](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/208)
+
+1. Support storing user data in Visual
+    * [Pull request 200](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/200)
+
+1. Add 16 bit format
+    * [Pull request 201](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/201)
+
+1. Destroy camera when the scene is destroyed
+    * [Pull request 175](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/175)
+
+1. Added function to get the number of channels from a GpuRay.
+    * [Pull request 178](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/178)
+
+### Ignition Rendering 2.X.X (2019-XX-XX)
+
+1. Backport setting cast shadows
+    * [Pull request 210](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/210)
+
+1. Ogre2: Add more shadow maps
+    * [Pull request 203](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/203)
+
+### Ignition Rendering 2.1.2 (2019-09-10)
+
+1. Enable gamma write on color texture in depth camera
+    * [Pull request 193](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/193)
+
+### Ignition Rendering 2.1.1 (2019-09-06)
+
+1. Add checks on mouse selection
+    * [Pull request 191](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/191)
+
+### Ignition Rendering 2.1.0 (2019-08-27)
+
+1. Support generating point cloud data in DepthCamera
+    * [Pull request 187](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/187)
+
+1. Fix GizmoVisual material memory leak
+    * [Pull request 185](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/185)
+
+1. Fix PixelFormat FLOAT32 byte and channel count
+    * [Pull request 186](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/186)
+
+### Ignition Rendering 2.0.1 (2019-07-16)
+
+1. Clean up camera objects when destroying scene
+    * [Pull request 174](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/174)
+    * [Pull request 180](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/180)
+
+### Ignition Rendering 2.0.0 (2019-05-20)
 
 1. More ogre2 error messages for debugging mesh loading.
     * [Pull request 170](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/170)